ZMETRIC-14: Refactored CumlativeFlowReportService into two files to separate out initialization functions.
