+++ src/CumulativeFlowReport/CumulativeFlowReportService.test.js
@@ -1,37 +1,7 @@
-import {CumulativeFlowReport} from "./CumulativeFlowReport";
-import {getColoursForNewIssues, mergeData} from "./CumulativeFlowReportService";
+import {mergeData} from "./CumulativeFlowReportService";
 
-describe("CumulativeFlowReport", () => {
-    const workflow = [
-        {
-            "id": "1",
-            "name": "Story",
-            "statuses": [
-                {
-                    "name": "To Do",
-                    "id": "1",
-                },
-                {
-                    "name": "In Progress",
-                    "id": "2"
-                },
-                {
-                    "name": "On Hold",
-                    "id": "3"
-                }
-            ]
-        }
-    ];
+describe("CumulativeFlowReportService", () => {
 
-    it("memorises the status colours", () => {
-        const expectedColours = new Map([
-            ["On Hold", "#ffc658"],
-            ["To Do", "#8884d8"],
-            ["In Progress", "#82ca9d"]
-        ]);
-
-        expect(getColoursForNewIssues(workflow)).toEqual(expectedColours);
-    });
 
     it('merges data flows for all selected issues', function () {
         const unmergedData = [
